Subscription

See merge request iseplife/api!59

+++ src/main/java/com/iseplife/api/services/FeedService.java
@@ -5,8 +5,10 @@ import com.iseplife.api.dao.feed.SubscriptionRepository;
 import com.iseplife.api.dto.view.PostView;
 import com.iseplife.api.entity.feed.Feed;
 import com.iseplife.api.entity.Subscription;
+import com.iseplife.api.entity.feed.Feedable;
 import com.iseplife.api.entity.user.Student;
 import com.iseplife.api.dao.feed.FeedRepository;
+import com.iseplife.api.exceptions.AuthException;
 import com.iseplife.api.exceptions.IllegalArgumentException;
 import com.iseplife.api.services.fileHandler.FileHandler;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -81,14 +83,23 @@ public class FeedService {
     return postService.getFeedDrafts(feed, author);
   }
 
-  public Boolean isSubscribed(Long id, Long studentID) {
-    return subscriptionRepository.findByFeedIdAndListenerId(id, studentID) != null;
+  public Boolean isSubscribedToFeed(Long id, Long studentID) {
+    return subscriptionRepository.existsSubscriptionByFeedIdAndListenerId(id, studentID);
   }
 
-  public void toggleSubscription(Long id, Long studentID) {
-    Subscription sub = subscriptionRepository.findByFeedIdAndListenerId(id, studentID);
+  public Boolean isSubscribedToFeed(Feedable feedable) {
+    return isSubscribedToFeed(feedable.getFeed().getId(), AuthService.getLoggedId());
+  }
+
+  public Boolean toggleSubscription(Long id, Long studentID) {
+    Feed feed = getFeed(id);
+    if(!AuthService.hasReadAccess(feed))
+      throw new AuthException("You have not sufficient rights on this feed (id:" + id + ")");
+
+    Subscription sub = subscriptionRepository.findByFeedIdAndListenerId(feed.getId(), studentID);
     if (sub != null) {
       subscriptionRepository.delete(sub);
+      return false;
     } else {
       Student student = studentService.getStudent(studentID);
       sub = new Subscription();
@@ -96,6 +107,7 @@ public class FeedService {
       sub.setListener(student);
 
       subscriptionRepository.save(sub);
+      return true;
     }
   }
 }
